Adding a debug output quad for the shadow map
Unfortunately with gDEBugger I can't interact with the scene, or even
run it on Linux for some reason, so this helps. I recall this not
working on my windows machine? Perhaps I fixed it now. Also migrating
the shadow map pass into its own function and running it each frame